♻️ Refactor: 공통 메서드 정리 및 코드 간결화
