+++ chessGame/src/chessGame/Pawn.java
@@ -140,123 +140,13 @@ public class Pawn extends ChessPiece {
 		}
 
 		// 공격
-		for (int i = 1; i <= 8; i++) {
-			for (int j = 1; j <= 8; j++) {
-				if (boards[i][j].getComponentCount() == 2) {
-					final int indexrow = i;
-					final int indexcol = j;
-					// 조건 : 현재 보드에 아이콘이 공격으로 바뀐 상대 말의 위치
-					if (((ImageIcon) ((ChessPiece) boards[i][j].getComponent(1)).getIcon()).getDescription()
-							.equals("white_icon_attack")) {
-
-						// 해당 액션리스너를 나중에 지우기 위해 따로 공격용 이벤트 변수를 만들어 저장
-						attackListener = new ActionListener() {
-							@Override
-							public void actionPerformed(ActionEvent e) {
-
-								int check = 0;
-								// 조건에 충족된 위치의 말이 체스말 모음에서 몇번 인덱스인지 확인 후 해당 말 비활성화
-								for (int z = 0; z < chesspiece_white.size(); z++) {
-									if (chesspiece_white
-											.get(z) == ((ChessPiece) boards[indexrow][indexcol].getComponent(1))) {
-										chesspiece_white.get(z).setEnabled(false);
-										check = z;
-									}
-								}
-
-								// 기준 위치에서 아군말 제거 후 이동할 위치에 상대말 제거 후 아군말 이동 위치로 추가
-								boards[row][col].remove(pawn);
-								boards[indexrow][indexcol].remove(chesspiece_white.get(check));
-								boards[indexrow][indexcol].add(pawn, "Center");
-
-								// 공격이 끝나면 모든 아군말을 일반 아이콘으로 전환
-								for (int k = 0; k < chesspiece_black.size(); k++) {
-									chesspiece_black.get(k).setIcon(chesspiece_black.get(k).black_icon);
-								}
-
-								// 무브핀 초기화
-								movepinsNotVisible();
-
-								// 현재 위치 값 저장
-								row = indexrow;
-								col = indexcol;
-
-								// 턴 정보를 상대 턴으로 변경
-								chessBoard.turn = "white";
-
-								p_board.getParent().validate();
-								p_board.getParent().repaint();
-
-								// 공격을 한 상태로 변경
-								attack = true;
-
-								// 사용 했거나 하지 않은 어택 리스너 제거
-								removeAttackBlack();
-
-							}
-						};
-
-						// 조건에 맞는 위치의 상대 말에 액션 이벤트 추가 후 어택 리스너 모음에 저장
-						((ChessPiece) boards[i][j].getComponent(1)).addActionListener(attackListener);
-						attackListeners.add(attackListener);
-
-					}
-				}
-
-			}
-
-		}
+		attackBlack(pawn);
 
 		// 무브핀의 중복 액션 이벤트를 막기 위해 이벤트 초기화
 		removeAction();
 
 		// 공격을 안했을 시 일반 이동 액션
-		if (!attack) {
-			for (int i = 1; i <= 8; i++) {
-				for (int j = 1; j <= 8; j++) {
-					final int indexrow = i;
-					final int indexcol = j;
-
-					// 조건 : 현재 보여지고 있는 무브핀
-					if (movepins[indexrow][indexcol].isVisible()) {
-						movepins[indexrow][indexcol].addActionListener(new ActionListener() {
-
-							@Override
-							public void actionPerformed(ActionEvent e) {
-
-								// 기준 위치에서 아군말 제거 후 이동할 위치에 아군 말 추가 후 일반 아이콘으로 변경
-								boards[row][col].remove(pawn);
-								boards[indexrow][indexcol].add(pawn, "Center");
-
-								// 이동이 끝나면 모든 아군말을 일반 아이콘으로 전환
-								for (int k = 0; k < chesspiece_black.size(); k++) {
-									chesspiece_black.get(k).setIcon(chesspiece_black.get(k).black_icon);
-								}
-
-								// 사용 안한 어택 리스너 제거
-								removeAttackBlack();
-
-								// 무브핀 초기화
-								movepinsNotVisible();
-
-								// 현재 위치 값 저장
-								row = indexrow;
-								col = indexcol;
-
-								// 턴 정보를 상대 턴으로 변경
-								chessBoard.turn = "white";
-
-								p_board.getParent().validate();
-								p_board.getParent().repaint();
-
-							}
-						});
-
-					}
-				}
-			}
-		}
-
+		moveBlack(pawn);
 	}
 
 	@Override
@@ -343,122 +233,13 @@ public class Pawn extends ChessPiece {
 		}
 
 		// 공격
-		for (int i = 1; i <= 8; i++) {
-			for (int j = 1; j <= 8; j++) {
-				if (boards[i][j].getComponentCount() == 2) {
-					final int indexrow = i;
-					final int indexcol = j;
-					// 조건 : 현재 보드에 아이콘이 공격으로 바뀐 상대 말의 위치
-					if (((ImageIcon) ((ChessPiece) boards[i][j].getComponent(1)).getIcon()).getDescription()
-							.equals("black_icon_attack")) {
-
-						// 해당 액션리스너를 나중에 지우기 위해 따로 공격용 이벤트 변수를 만들어 저장
-						attackListener = new ActionListener() {
-							@Override
-							public void actionPerformed(ActionEvent e) {
-
-								int check = 0;
-								// 조건에 충족된 위치의 말이 체스말 모음에서 몇번 인덱스인지 확인 후 해당 말 비활성화
-								for (int z = 0; z < chesspiece_black.size(); z++) {
-									if (chesspiece_black
-											.get(z) == ((ChessPiece) boards[indexrow][indexcol].getComponent(1))) {
-										chesspiece_black.get(z).setEnabled(false);
-										check = z;
-									}
-								}
-
-								// 기준 위치에서 아군말 제거 후 이동할 위치에 상대말 제거 후 아군말 이동 위치로 추가
-								boards[row][col].remove(pawn);
-								boards[indexrow][indexcol].remove(chesspiece_black.get(check));
-								boards[indexrow][indexcol].add(pawn, "Center");
-
-								// 공격이 끝나면 모든 아군말을 일반 아이콘으로 전환
-								for (int k = 0; k < chesspiece_white.size(); k++) {
-									chesspiece_white.get(k).setIcon(chesspiece_white.get(k).white_icon);
-								}
-
-								// 무브핀 초기화
-								movepinsNotVisible();
-
-								// 현재 위치 값 저장
-								row = indexrow;
-								col = indexcol;
-
-								// 턴 정보를 상대 턴으로 변경
-								chessBoard.turn = "black";
-
-								p_board.getParent().validate();
-								p_board.getParent().repaint();
-
-								// 공격을 한 상태로 변경
-								attack = true;
-
-								// 사용 했거나 하지 않은 어택 리스너 제거
-								removeAttackBlack();
-
-							}
-						};
-
-						// 조건에 맞는 위치의 상대 말에 액션 이벤트 추가 후 어택 리스너 모음에 저장
-						((ChessPiece) boards[i][j].getComponent(1)).addActionListener(attackListener);
-						attackListeners.add(attackListener);
-
-					}
-				}
-
-			}
-
-		}
+		attackWhite(pawn);
 
 		// 무브핀의 중복 액션 이벤트를 막기 위해 이벤트 초기화
 		removeAction();
 
 		// 공격을 안했을 시 일반 이동 액션
-		if (!attack) {
-			for (int i = 1; i <= 8; i++) {
-				for (int j = 1; j <= 8; j++) {
-					final int indexrow = i;
-					final int indexcol = j;
-
-					// 조건 : 현재 보여지고 있는 무브핀
-					if (movepins[indexrow][indexcol].isVisible()) {
-						movepins[indexrow][indexcol].addActionListener(new ActionListener() {
-
-							@Override
-							public void actionPerformed(ActionEvent e) {
-
-								// 기준 위치에서 아군말 제거 후 이동할 위치에 아군 말 추가 후 일반 아이콘으로 변경
-								boards[row][col].remove(pawn);
-								boards[indexrow][indexcol].add(pawn, "Center");
-
-								// 이동이 끝나면 모든 아군말을 일반 아이콘으로 전환
-								for (int k = 0; k < chesspiece_white.size(); k++) {
-									chesspiece_white.get(k).setIcon(chesspiece_white.get(k).white_icon);
-								}
-
-								// 사용 안한 어택 리스너 제거
-								removeAttackBlack();
-
-								// 무브핀 초기화
-								movepinsNotVisible();
-
-								// 현재 위치 값 저장
-								row = indexrow;
-								col = indexcol;
-
-								// 턴 정보를 상대 턴으로 변경
-								chessBoard.turn = "black";
-
-								p_board.getParent().validate();
-								p_board.getParent().repaint();
-
-							}
-						});
-
-					}
-				}
-			}
-		}
+		moveWhite(pawn);
 
 	}
 }
